fix previous commit: add hmac doc and remove rebar
The previous commit added the hmac module and was supposed to include
hmac documentation, but by accident it didn't. Add hmac specs and
doc.

Also remove local copy of rebar -- rebar is popular enough now that
erlang developers generally need a copy on their system, and we don't
want to have to keep updating a local copy here every time the rebar
devs make a change.