replace hmac:hexlify with a NIF
In pull request 9, @bosky101 showed that the performance of hmac:hexlify/1
could be increased in pure Erlang code, but my further experimentation
showed that even better performance could be achieved by replacing
hmac:hexlify/1 with a NIF, an approach that fits well with erlsha2's
existing use of NIFs.

Also add hmac:hexlify/2 to allow the caller to specify the function's
return value as either a string or a binary and with either uppercase or
lowercase alphabetic hex digits. For backward compatibility, the default is
to return an uppercase string, same as hmac:hexlify/1.

Add tests for hmac:hexlify/2 and update the docs.

Update version to 2.1.

Many thanks to @bosky101 for finding this performance issue and also for
suggesting adding support for both string and binary hexlify return values.